refactor: :lipstick: update ui for layout
