+++ assets/css/main.css
@@ -1,28 +1,23 @@
 @import "tailwindcss";
 @import "@nuxt/ui";
 
-@theme static {
+@theme extend {
   --breakpoint-xs: 450px;
   --breakpoint-xxs: 360px;
+
+  --ui-color-neutral-75: #f4f7fa;
+  --ui-color-neutral-1000: #01040b;
 }
 
-/** TODO: update overall theme 
-* Hue 'primary', 'secondary' etc. is set in app.config. Here we set the shade (500)
-* Find a way to handle dark mode
-*/
 :root {
-  --ui-primary: var(--ui-color-primary-500);
-  --ui-secondary: var(--ui-color-secondary-500);
-  --ui-success: var(--ui-color-success-500);
-  --ui-info: var(--ui-color-info-500);
-  --ui-warning: var(--ui-color-warning-500);
-  --ui-error: var(--ui-color-error-500);
+  --ui-bg: var(--ui-color-neutral-50);
+  --ui-border: var(--ui-color-neutral-300);
 }
 .dark {
-  --ui-primary: var(--ui-color-primary-400);
-  --ui-secondary: var(--ui-color-secondary-400);
-  --ui-success: var(--ui-color-success-400);
-  --ui-info: var(--ui-color-info-400);
-  --ui-warning: var(--ui-color-warning-400);
-  --ui-error: var(--ui-color-error-400);
+  --ui-bg: var(--ui-color-neutral-950);
+  --ui-border: var(--color-neutral-600);
+}
+
+.blurry-card {
+  @apply bg-neutral-100/80 dark:bg-neutral-800/80 border border-neutral-300 dark:border-neutral-600 rounded-md backdrop-blur-lg transition-transform ease-in-out duration-500;
 }
